Перенес сообщения в Redux
