+++ package.json
@@ -17,6 +17,7 @@
     "material-ui": "^0.20.2",
     "prop-types": "^15.7.2",
     "react": "^16.10.2",
+    "react-addons-update": "^15.6.2",
     "react-dom": "^16.10.2",
     "react-redux": "^7.1.1",
     "react-router-dom": "^5.1.2",
