--- static_src/Components/Message.jsx
@@ -1,24 +1,0 @@
-
-import React from 'react';
-
-
-
-
-export default class Message extends React.Component {
-	constructor(props) { 
-    super (props) 
-	}
-	
-	render() {
-		return <div
-		className="message"
-		style={ { alignSelf: this.props.sender === 'bot' ?'flex-start' : 'flex-end' } }>
-		<div>{ this.props.text }</div>
-		<div className="message-sender">{ this.props.sender }</div>
-		
-		</div>
-
-    }
-}
-
-
