+++ static_src/actions/chatActions.js
@@ -1,0 +1,6 @@
+export const ADD_CHAT = '@@chat/ADD_CHAT';
+
+export const addChat = (title) => ({
+   type: ADD_CHAT,
+   title,
+});
